# frontend/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Physionexs API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/me/care-plans": {
      "get": {
        "tags": [
          "care"
        ],
        "summary": "My Plans",
        "operationId": "my_plans_me_care_plans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CarePlanOut"
                  },
                  "type": "array",
                  "title": "Response My Plans Me Care Plans Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/me/today": {
      "get": {
        "tags": [
          "care"
        ],
        "summary": "Today",
        "operationId": "today_me_today_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "day",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/exercise-logs": {
      "post": {
        "tags": [
          "care"
        ],
        "summary": "Log Exercise",
        "operationId": "log_exercise_me_exercise_logs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseLogIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseLogOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/me/exercise-logs/{plan_exercise_id}": {
      "delete": {
        "tags": [
          "care"
        ],
        "summary": "Undo Exercise",
        "description": "Undo the latest completion (for mis-taps). Points already awarded are kept.",
        "operationId": "undo_exercise_me_exercise_logs__plan_exercise_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Plan Exercise Id"
            }
          },
          {
            "name": "logged_on",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Logged On"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/dose-logs": {
      "post": {
        "tags": [
          "care"
        ],
        "summary": "Log Dose",
        "operationId": "log_dose_me_dose_logs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoseLogIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Log Dose Me Dose Logs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "care"
        ],
        "summary": "Undo Dose",
        "operationId": "undo_dose_me_dose_logs_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "medication_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Medication Id"
            }
          },
          {
            "name": "logged_on",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Logged On"
            }
          },
          {
            "name": "dose_slot",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dose Slot"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/progress": {
      "get": {
        "tags": [
          "care"
        ],
        "summary": "Progress",
        "operationId": "progress_me_progress_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "week",
                "month"
              ],
              "type": "string",
              "default": "week",
              "title": "Range"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/prescriptions": {
      "get": {
        "tags": [
          "care"
        ],
        "summary": "My Prescriptions",
        "operationId": "my_prescriptions_me_prescriptions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PrescriptionOut"
                  },
                  "type": "array",
                  "title": "Response My Prescriptions Me Prescriptions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CarePlanOut": {
        "properties": {
          "condition": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Condition"
          },
          "condition_detail": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition Detail"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "sessions_planned": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 200.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sessions Planned"
          },
          "ends_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends On"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/CarePlanStatus"
          },
          "starts_on": {
            "type": "string",
            "format": "date",
            "title": "Starts On"
          },
          "physio_name": {
            "type": "string",
            "title": "Physio Name"
          },
          "clinic_name": {
            "type": "string",
            "title": "Clinic Name"
          },
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/PlanExerciseOut"
            },
            "type": "array",
            "title": "Exercises"
          },
          "medications": {
            "items": {
              "$ref": "#/components/schemas/MedicationOut"
            },
            "type": "array",
            "title": "Medications"
          },
          "tests": {
            "items": {
              "$ref": "#/components/schemas/TestOrderOut"
            },
            "type": "array",
            "title": "Tests"
          }
        },
        "type": "object",
        "required": [
          "condition",
          "id",
          "status",
          "starts_on",
          "physio_name",
          "clinic_name",
          "exercises",
          "medications",
          "tests"
        ],
        "title": "CarePlanOut"
      },
      "CarePlanStatus": {
        "type": "string",
        "enum": [
          "active",
          "completed",
          "cancelled"
        ],
        "title": "CarePlanStatus"
      },
...
      "DoseLogIn": {
        "properties": {
          "medication_id": {
            "type": "string",
            "format": "uuid",
            "title": "Medication Id"
          },
          "logged_on": {
            "type": "string",
            "format": "date",
            "title": "Logged On"
          },
          "dose_slot": {
            "type": "string",
            "pattern": "^\\d{2}:\\d{2}$",
            "title": "Dose Slot"
          }
        },
        "type": "object",
        "required": [
          "medication_id",
          "logged_on",
          "dose_slot"
        ],
        "title": "DoseLogIn"
      },
      "DoseUnit": {
        "type": "string",
        "enum": [
          "reps",
          "seconds"
        ],
        "title": "DoseUnit"
      },
...
      "ExerciseLogIn": {
        "properties": {
          "plan_exercise_id": {
            "type": "string",
            "format": "uuid",
            "title": "Plan Exercise Id"
          },
          "logged_on": {
            "type": "string",
            "format": "date",
            "title": "Logged On"
          },
          "feel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Feel"
              },
              {
                "type": "null"
              }
            ]
          },
          "pain": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Pain"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "plan_exercise_id",
          "logged_on"
        ],
        "title": "ExerciseLogIn"
      },
      "ExerciseLogOut": {
        "properties": {
          "done": {
            "type": "integer",
            "title": "Done"
          },
          "scheduled": {
            "type": "integer",
            "title": "Scheduled"
          },
          "streak_days": {
            "type": "integer",
            "title": "Streak Days"
          },
          "points_awarded": {
            "type": "integer",
            "title": "Points Awarded"
          }
        },
        "type": "object",
        "required": [
          "done",
          "scheduled",
          "streak_days",
          "points_awarded"
        ],
        "title": "ExerciseLogOut"
      },
...
      "Feel": {
        "type": "string",
        "enum": [
          "easy",
          "ok",
          "hard"
        ],
        "title": "Feel"
      },
      "Frequency": {
        "type": "string",
        "enum": [
          "daily",
          "alternate_days",
          "weekly_3x",
          "weekly"
        ],
        "title": "Frequency"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanExerciseOut": {
        "properties": {
          "exercise_id": {
            "type": "string",
            "format": "uuid",
            "title": "Exercise Id"
          },
          "sets": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Sets"
          },
          "reps": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 200.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Reps"
          },
          "hold_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 600.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Hold Seconds"
          },
          "rest_seconds": {
            "type": "integer",
            "maximum": 600.0,
            "minimum": 0.0,
            "title": "Rest Seconds",
            "default": 30
          },
          "frequency": {
            "$ref": "#/components/schemas/Frequency",
            "default": "daily"
          },
          "times_per_day": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "Times Per Day",
            "default": 1
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "body_region": {
            "type": "string",
            "title": "Body Region"
          },
          "dose_unit": {
            "$ref": "#/components/schemas/DoseUnit"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          }
        },
        "type": "object",
        "required": [
          "exercise_id",
          "sets",
          "id",
          "name",
          "body_region",
          "dose_unit"
        ],
        "title": "PlanExerciseOut"
      },
...
      "PrescriptionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "rx_no": {
            "type": "string",
            "title": "Rx No"
          },
          "issued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Issued At"
          },
          "snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Snapshot"
          }
        },
        "type": "object",
        "required": [
          "id",
          "rx_no",
          "issued_at",
          "snapshot"
        ],
        "title": "PrescriptionOut"
      },
...
      "ProgressOut": {
        "properties": {
          "days": {
            "items": {
              "$ref": "#/components/schemas/ProgressDay"
            },
            "type": "array",
            "title": "Days"
          },
          "adherence_pct": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adherence Pct"
          },
          "streak_days": {
            "type": "integer",
            "title": "Streak Days"
          },
          "best_streak_days": {
            "type": "integer",
            "title": "Best Streak Days"
          },
          "pain_now": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pain Now"
          },
          "pain_start": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pain Start"
          },
          "unlogged_doses_today": {
            "type": "integer",
            "title": "Unlogged Doses Today"
          }
        },
        "type": "object",
        "required": [
          "days",
          "adherence_pct",
          "streak_days",
          "best_streak_days",
          "pain_now",
          "pain_start",
          "unlogged_doses_today"
        ],
        "title": "ProgressOut"
      },
...
      "TestOrderOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "$ref": "#/components/schemas/TestOrderStatus"
          },
          "result_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Note"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "result_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "result_note",
          "created_at",
          "result_at"
        ],
        "title": "TestOrderOut"
      },
      "TestOrderStatus": {
        "type": "string",
        "enum": [
          "ordered",
          "result_ready",
          "cancelled"
        ],
        "title": "TestOrderStatus"
      },
...
      "TodayOut": {
        "properties": {
          "day": {
            "type": "string",
            "format": "date",
            "title": "Day"
          },
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/TodayExercise"
            },
            "type": "array",
            "title": "Exercises"
          },
          "doses": {
            "items": {
              "$ref": "#/components/schemas/TodayDose"
            },
            "type": "array",
            "title": "Doses"
          },
          "exercise_pct": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exercise Pct"
          },
          "dose_pct": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dose Pct"
          },
          "streak_days": {
            "type": "integer",
            "title": "Streak Days"
          }
        },
        "type": "object",
        "required": [
          "day",
          "exercises",
          "doses",
          "exercise_pct",
          "dose_pct",
          "streak_days"
        ],
        "title": "TodayOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}